dev/rag/requirements.txt
google-generativeai>=0.3.0
chromadb>=0.4.0
langchain>=0.1.0
python-dotenv>=1.0.0